+++ .idea/misc.xml
@@ -50,6 +50,8 @@
         <entry key="..\:/Users/marti/Documents/AndroidStudioProjects/our-planner/app/src/main/res/layout/activity_view_event.xml" value="0.1" />
         <entry key="..\:/Users/marti/Documents/AndroidStudioProjects/our-planner/app/src/main/res/layout/fragment_calendar.xml" value="0.17552083333333332" />
         <entry key="..\:/Users/marti/Documents/AndroidStudioProjects/our-planner/app/src/main/res/layout/fragment_invitations.xml" value="0.17552083333333332" />
+        <entry key="..\:/Users/marti/Documents/AndroidStudioProjects/our-planner/app/src/main/res/layout/fragment_month_calendar.xml" value="0.17552083333333332" />
+        <entry key="..\:/Users/marti/Documents/AndroidStudioProjects/our-planner/app/src/main/res/layout/fragment_week_calendar.xml" value="0.17552083333333332" />
         <entry key="..\:/Users/marti/Documents/AndroidStudioProjects/our-planner/app/src/main/res/layout/groups_spinner.xml" value="0.17552083333333332" />
         <entry key="..\:/Users/marti/Documents/AndroidStudioProjects/our-planner/app/src/main/res/layout/item_list_calendar_groups.xml" value="0.17552083333333332" />
         <entry key="..\:/Users/marti/Documents/AndroidStudioProjects/our-planner/app/src/main/res/layout/item_list_comments.xml" value="0.17552083333333332" />
